api/requirements.txt
werkzeug==2.3.8
flask==2.3.3
flask-restful==0.3.9
flask-login==0.6.2
flask-mysql==1.5.2
cryptography==38.0.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
GeoAlchemy2==0.16.0
PyMySQL==1.1.1
python-dotenv==1.0.1
Flask-JWT-Extended==4.7.1
shapely==2.0.6
geojson==3.2.0
Flask-SQLAlchemy==3.1.1